refactor: create prune package
